Enable multi-node conformance testing
