+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/LedgerApiTestTool.scala
@@ -93,12 +93,10 @@ object LedgerApiTestTool {
       })
 
     val runner = new LedgerTestSuiteRunner(
-      Vector(
-        LedgerSessionConfiguration(
-          config.host,
-          config.port,
-          config.tlsConfig,
-          config.commandSubmissionTtlScaleFactor)),
+      LedgerSessionConfiguration(
+        config.participants,
+        config.tlsConfig,
+        config.commandSubmissionTtlScaleFactor),
       testsToRun.values.toVector,
       config.timeoutScaleFactor,
       identifierSuffix
