+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/infrastructure/participant/ParticipantSessionConfiguration.scala
@@ -1,11 +1,12 @@
 // Copyright (c) 2019 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.api.testtool.infrastructure
+package com.daml.ledger.api.testtool.infrastructure.participant
 
 import com.digitalasset.ledger.api.tls.TlsConfiguration
 
-private[testtool] final case class LedgerSessionConfiguration(
-    participants: Vector[(String, Int)],
+private[testtool] final case class ParticipantSessionConfiguration(
+    host: String,
+    port: Int,
     ssl: Option[TlsConfiguration],
     commandTtlFactor: Double)
