+++ ledger/ledger-api-test-tool/src/main/scala/com/daml/ledger/api/testtool/tests/Identity.scala
@@ -5,14 +5,16 @@ package com.daml.ledger.api.testtool.tests
 
 import com.daml.ledger.api.testtool.infrastructure.{LedgerSession, LedgerTest, LedgerTestSuite}
 
-import scala.concurrent.Future
-
 final class Identity(session: LedgerSession) extends LedgerTestSuite(session) {
 
   private[this] val identity =
     LedgerTest("IdNotEmpty", "A ledger should return a non-empty string as its identity") {
-      ledger =>
-        Future(assert(ledger.ledgerId.nonEmpty, "The returned ledger identifier was empty"))
+      context =>
+        for {
+          ledger <- context.participant()
+        } yield {
+          assert(ledger.ledgerId.nonEmpty, "The returned ledger identifier was empty")
+        }
     }
 
   override val tests: Vector[LedgerTest] = Vector(identity)
